MGS-3522: Fix termios issue when running in batch mode.
We disable input buffering using termios/tty methods. This also
takes care of not aggressively reading counter values by having
a time-out delay in pselect().

When used in a script and put in bg, tty incorrectly considers that we're
receiving something from the keyboard. Luckily we can detect when this happens
and for the user to use a special switch which inserts artificially a time-out
of 1 second.

Signed-off-by: Marius Vlad <marius-cristian.vlad@nxp.com>